move dependencies to root level make it available to CLICommand in typed form
