+++ resotocore/tests/resotocore/config/core_config_handler_test.py
@@ -14,10 +14,9 @@ from resotocore.core_config import (
     CustomCommandsConfig,
     ResotoCoreSnapshotsRoot,
 )
-from resotocore.dependencies import empty_config
+from resotocore.system_start import empty_config
 from resotocore.message_bus import MessageBus, CoreMessage
 from resotocore.ids import ConfigId
-from resotocore.types import Json
 
 
 @mark.asyncio
